Modifications to resolving cinder device names
When listing all volumes, return the device name reported by cinder. Don't
try to resolve local device names.

When attaching volume, return error if there is an in error resolving device
name (by-id link does not exist). Don't ever return the device name reported
by Cinder.

When inspecting volume, try to resolve local device name and return 'unknown'
if there is an error in resolving the name (by-id link does not exist). Don't
ever return the device name reported by Cinder.

This change is done to make sure that wrong volumes are not attached to
containers, which can happen if by-id links are missing and Cinder device names
(that happen to be wrong) are used as a fallback.

Returning "unknown" as device name should cause an error when trying to mount
the volume to container and lead to restarting the container and then the volume
attachment is also retried.

It is not enough to change this only in attach because inspect is also called
when mounting the volume to container.

Returning device name 'unknown' can then happen in two cases:

1. Inspecting volume that is attached on the instance where this code is
executed and there is an error resolving the device name (the by-id link does
not exist)

2. Inspecting volume that is attached to some other instance than where this
code is executed

Returning "unknown" in case 1 should ensure that a wrong device is not attached
to container in any case.

This is a workaround but maybe it is ok considering that this was not the best
place to fix the problem with Cinder device names in the first place because it
does not work when rexray is in client-server mode (this code is executed on
the server and not on the instance that is attaching / inspecting the volume).